+++ CHANGELOG.md
@@ -41,4 +41,4 @@
 
 - Added latestFileUpload to dataSource endpoint
 - Added fetch to dataSource endpoint
-
+- Added update to dataSource endpoint